OK-747: Refaktoroitu datan uudelleenlatausta valintalaskennan tulosten muokkauksen jälkeen
- Henkilöittäin näkymässä noudetaan eri rajapinnalla kuin hakukohteen valintalaskennan-tulokset-välilehdellä. Annetaan onSuccess-parametri, jonka avulla noudetaan uudelleen käyttäen oikeaa rajapintaa kummassakin tapauksessa.
