+++ src/app/haku/[oid]/henkilo/[hakemusOid]/components/hakutoive-accordion-content.tsx
@@ -14,6 +14,8 @@ import { ValinnanTulosCells } from './valinnan-tulos-cells';
 import { styled } from '@/app/lib/theme';
 import { EditButton } from '@/app/components/edit-button';
 import { HenkilonHakukohdeTuloksilla } from '../lib/henkilo-page-types';
+import { hakemuksenValintalaskennanTuloksetQueryOptions } from '@/app/lib/valintalaskenta-service';
+import { QueryClient, useQueryClient } from '@tanstack/react-query';
 
 const HakutoiveInfoRow = styled(TableRow)({
   '&:nth-of-type(even)': {
@@ -23,6 +25,23 @@ const HakutoiveInfoRow = styled(TableRow)({
     backgroundColor: ophColors.lightBlue2,
   },
 });
+
+const refetchValinnanvaiheet = ({
+  queryClient,
+  hakuOid,
+  hakemusOid,
+}: {
+  queryClient: QueryClient;
+  hakuOid: string;
+  hakemusOid: string;
+}) => {
+  const options = hakemuksenValintalaskennanTuloksetQueryOptions({
+    hakuOid,
+    hakemusOid,
+  });
+  queryClient.resetQueries(options);
+  queryClient.invalidateQueries(options);
+};
 
 export const HakutoiveAccordionContent = ({
   hakija,
@@ -35,6 +54,8 @@ export const HakutoiveAccordionContent = ({
 }) => {
   const { t } = useTranslations();
   const { valinnanvaiheet } = hakukohde;
+
+  const queryClient = useQueryClient();
 
   return isEmpty(valinnanvaiheet ?? []) ? (
     <HakutoiveInfoRow>
@@ -78,6 +99,13 @@ export const HakutoiveAccordionContent = ({
                     hakukohde,
                     valintatapajono: jono,
                     jonosija: jono.jonosijat?.[0], // Yhdellä henkilöllä vain yksi jonosija
+                    onSuccess: () => {
+                      refetchValinnanvaiheet({
+                        hakuOid: hakukohde.hakuOid,
+                        hakemusOid: jonosija.hakemusOid,
+                        queryClient,
+                      });
+                    },
                   })
                 }
               />
